requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.4.4
certifi==2026.2.25
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==6.0.2
django-autocomplete-light==3.12.1
django-autoslug==1.9.9
django-browser-reload==1.21.0
django-cities-light==3.11
django-phonenumber-field==8.4.0
django-tailwind==4.4.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
phonenumbers==9.0.25
progressbar2==4.5.0
psutil==7.2.2
psycopg2==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
